Add Link to Use Local Database
Previously an unused option existed in the code to use H2, an in-memory
database for the DB operations. This was never exposed anywhere, and
still required users to "sign in".

This commit adds a new link (and corresponding method) to use H2 instead
of signing in, modifying the header. All functionality should otherwise
be identical.